ToDoList/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
pytz==2021.3
selenium==3.141.0
sqlparse==0.4.2
urllib3==1.26.7
